requirements.txt
streamlit
polars
plotly
unidecode
python-dateutil